third_party/WebKit/LayoutTests/css3/device-adapt/opera/constrain-010-expected.txt


FAIL CSS Test: @viewport constrained - width 20rem assert_equals: expected 320 but got 800
